requirements.txt
numpy
robotbackgroundlogger
robotframework
configparser
requests
robotframework-requests
RESTinstance
PyYAML
psutil